+++ changelog.md
@@ -4,3 +4,8 @@
 - Add the color and decoration setter
 - Implement the radio choice for the Text User Interface
 
+## Version 0.2
+- Add checkboxes method, return array of byte indicate the state of that current option
+- Update sample.c file (simple sort method panel);
+- Fix the display bug
+